Rename "sdb.repl" to "sdb.internal.repl"
This renames the "sdb.repl" modules to be within "sdb.internal" since
the REPL used by sdb is meant to only be used internally by "sdb".
